+++ js/util.js
@@ -1,0 +1,21 @@
+const getRandomNumber = (min , max) => {
+  if (!(typeof min === 'number') || !(typeof max === 'number')) {
+    return 'Введите число';
+  }
+  if (min < 0 || max <= 0 || max < min) {
+    return 'Неверный диапазон чисел'
+  }
+  return Math.floor(Math.random() * (max - min + 1) + min);
+};
+
+const getFloatRandomNumber = (min, max, point) => {
+  if (!(typeof min === 'number') || !(typeof max === 'number') || !(typeof point === 'number')) {
+    return 'Введите число';
+  }
+  if (min < 0 || max <= 0 || max < min) {
+    return 'Неверный диапазон чисел'
+  }
+  return +(Math.random() * (max - min) + min).toFixed(point);
+};
+
+export {getRandomNumber, getFloatRandomNumber};
